Move implicits to the companion object
