+++ core/src/main/scala/neotypes/implicits/mappers/all.scala
@@ -1,8 +1,4 @@
 package neotypes
 package implicits.mappers
 
-trait AllMappers
-  extends ExecutionMappers
-  with ParameterMappers
-  with ResultMappers
-  with CompatibilityMappers
+trait AllMappers extends CompatibilityMappers
