+++ core/src/main/scala/neotypes/implicits/mappers/package.scala
@@ -2,8 +2,5 @@ package neotypes.implicits
 
 package object mappers {
   final object all extends AllMappers
-  final object executions extends ExecutionMappers
-  final object parameters extends ParameterMappers
-  final object results extends ResultMappers with CompatibilityMappers
-  final object values extends ValueMappers
+  final object collections extends CompatibilityMappers
 }
