+++ core/src/test/scala/neotypes/PathSessionSpec.scala
@@ -1,7 +1,6 @@
 package neotypes
 
 import neotypes.generic.auto._
-import neotypes.implicits.mappers.results._
 import neotypes.implicits.syntax.string._
 import neotypes.internal.syntax.async._
 import org.neo4j.driver.types.{Node, Relationship}
