+++ site/src/main/mdoc/overview.md
@@ -40,7 +40,6 @@ Or, if you use other effect types instead of `Future`, for example, `IO`. Then, 
 ```scala mdoc:compile-only
 import neotypes.Driver
 import neotypes.generic.auto._ // Allows to automatically derive an implicit ResultMapper for case classes.
-import neotypes.implicits.mappers.results._ // Provides instances for primitives (String, Int, Long, Instant, etc)
 import neotypes.implicits.syntax.string._ // Provides the query[T] extension method.
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.Implicits.global
@@ -58,10 +57,10 @@ def result(driver: Driver[Future]): Future[List[Movie]] =
 ## Query execution
 
 Once you have a `Session` constructed, you can start querying the database.
-The import `neotypes.implicits.all._` adds an extension method `query[T]` to each string literal in its scope, or you can use the cypher _(`c`)_ string interpolator.
+The import `neotypes.implicits.syntax.all._` adds an extension method `query[T]` to each string literal in its scope, or you can use the cypher _(`c`)_ string interpolator.
 
 ```scala mdoc:invisible
-import neotypes.implicits.all._
+import neotypes.implicits.syntax.all._
 def session: neotypes.Session[scala.concurrent.Future] = ???
 ```
 
@@ -90,12 +89,13 @@ If you need to support your return types for this type of queries, you can provi
 * `set(session)` - runs a query and returns a **Set** of results.
 * `vector(session)` - runs a query and returns a **Vector** of results.
 * `map(session)` - runs a query and returns a **Map** of results _(only if the elements are tuples)_.
-* `collectAs(Col)(session)` - runs a query and retunrs a **Col** of results _(where **Col** is any kind of collection)_.
+* `collectAs(Col)(session)` - runs a query and returns a **Col** of results _(where **Col** is any kind of collection)_. Might require the following import `import neotypes.implicits.mappers.collections._`.
 * `stream(session)` - runs a query and returns a **Stream** of results
 _(for more information, please read [streaming](streams))_.
 
 ```scala mdoc:compile-only
 import neotypes.generic.auto._
+import neotypes.implicits.mappers.collections._
 import org.neo4j.driver.Value
 import shapeless.{::, HNil}
 import scala.collection.immutable.{ListMap, ListSet}
